+++ package.json
@@ -4,15 +4,17 @@
   "private": true,
   "description": "Учебный проект «Менеджер задач» от HTML Academy",
   "scripts": {
-    "build" : "webpack",
+    "build": "webpack",
     "eslint": "eslint src/**/*.js",
-    "test": "npm run eslint"
+    "test": "npm run eslint",
+    "start": "webpack-dev-server --open"
   },
   "devDependencies": {
     "eslint": "6.1.0",
     "eslint-config-htmlacademy": "0.5.2",
     "webpack": "4.39.2",
-    "webpack-cli": "3.3.7"
+    "webpack-cli": "3.3.7",
+    "webpack-dev-server": "3.8.0"
   },
   "engines": {
     "node": "10"
