+++ public/index.html
@@ -53,6 +53,6 @@
       </section>
 
     </main>
-    <script src="../src/main.js"></script>
+    <script type="module" src="bundle.js"></script>
   </body>
 </html>
