+++ webpack.config.js
@@ -1,11 +1,18 @@
-const path = require("path");
+/* eslint-disable no-undef */
+const path = require(`path`);
 
 module.exports = {
-    mode: `development`,
-    entry: `./src/main.js`,
-    output: {
-        filename: `bundle.js`,
-        path: path.join(__dirname, `public`),
-        devtool: `source-map`
-    }
-}
+  mode: `development`,
+  entry: `./src/main.js`,
+  output: {
+    filename: `bundle.js`,
+    path: path.join(__dirname, `public`),
+  },
+  devtool: `source-map`,
+  devServer: {
+    contentBase: path.join(__dirname, `public`),
+    publicPath: `http://localhost:8080/`,
+    compress: true,
+    watchContentBase: true
+  }
+};
